Add modified center pasting
