+++ amodal_utils.py
@@ -1,107 +1,8 @@
 import json
-import os
+import os, sys
 import cv2
 import numpy as np
-
-def get_image_paths_from_folder(folder_path, extensions=['.jpg', '.png']):
-    image_paths = []
-    for filename in os.listdir(folder_path):
-        if any(filename.endswith(ext) for ext in extensions):
-            image_paths.append(os.path.join(folder_path, filename))
-    return image_paths
-
-def save_image(save_dir, file_name, img, flag=0):
-    key = 'Mask' if flag == 0 else 'Image'
-
-    os.makedirs(save_dir, exist_ok=True)  # 저장 경로가 없으면 생성
-    image_save_path = os.path.join(save_dir, file_name)
-    cv2.imwrite(image_save_path, img)
-    print(f"{key} 저장됨: {image_save_path}")
-    return image_save_path
-
-def get_bbox_from_mask(mask):
-    """
-    마스크에서 일반적인 [x_min, y_min, x_max, y_max] 형식의 bbox 추출
-    """
-    coords = np.column_stack(np.where(mask == 255))  # 마스크 좌표 추출
-    if coords.size == 0:  # 빈 마스크 처리
-        return [0, 0, 0, 0]
-    y_min, x_min = coords.min(axis=0)
-    y_max, x_max = coords.max(axis=0)
-    return [x_min, y_min, x_max, y_max]
-
-def get_coco_bbox_from_mask(mask):
-    """
-    마스크에서 COCO 포맷 [x_min, y_min, width, height] 형식의 bbox 추출
-    """
-    x_min, y_min, x_max, y_max = get_bbox_from_mask(mask)
-    return [float(x_min), float(y_min), float(x_max - x_min), float(y_max - y_min)]
-
-def mask_to_polygon(binary_mask, min_area=60):
-    # Find contours
-    contours, _ = cv2.findContours(binary_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-    polygons = []
-
-    for contour in contours:
-        # Simplify the contour to reduce the number of points
-        epsilon = 0.01 * cv2.arcLength(contour, True)
-        contour = cv2.approxPolyDP(contour, epsilon, True).flatten().tolist()
-
-        # Check if the contour is valid (minimum points and area)
-        if len(contour) > 4 and cv2.contourArea(np.array(contour).reshape(-1, 2)) > min_area:
-            polygons.append(contour)
-    
-    return polygons
-
-def mask_to_polygon(binary_mask, min_contour_area=10):
-    """
-    Convert a binary mask to a COCO segmentation polygon, filtering by area.
-    """
-    contours, _ = cv2.findContours(binary_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-    polygons = []
-    for contour in contours:
-        if cv2.contourArea(contour) >= min_contour_area:  # 최소 크기 필터링
-            contour = contour.flatten().tolist()
-            if len(contour) > 4:  # 최소한의 폴리곤 점이 필요
-                polygons.append(contour)
-    return polygons
-
-
-def resize_image_and_masks(image, masks, target_size=(768, 1024)):
-    target_width, target_height = target_size
-    original_height, original_width = image.shape[:2]
-    
-    # 이미지 리사이즈
-    resized_image = cv2.resize(image, (target_width, target_height), interpolation=cv2.INTER_LINEAR)
-    
-    # 마스크 리사이즈
-    resized_masks = [
-        cv2.resize(mask, (target_width, target_height), interpolation=cv2.INTER_NEAREST)
-        for mask in masks
-    ]
-    
-    # 스케일 계산
-    scale_x = target_width / original_width
-    scale_y = target_height / original_height
-    
-    return resized_image, resized_masks, scale_x, scale_y
-
-
-
-def resize_leaf_image(leaf_image, cucumber_image):
-    cucumber_h, cucumber_w = cucumber_image.shape[:2]
-    leaf_h, leaf_w = leaf_image.shape[:2]
-
-    # 잎 이미지가 오이 이미지를 초과하면 비율에 맞춰 리사이즈
-    if leaf_h > cucumber_h or leaf_w > cucumber_w:
-        scale = min(cucumber_h / leaf_h, cucumber_w / leaf_w)  # 가장 작은 스케일로 맞춤
-        new_leaf_h = int(leaf_h * scale)
-        new_leaf_w = int(leaf_w * scale)
-        leaf_image = cv2.resize(leaf_image, (new_leaf_w, new_leaf_h), interpolation=cv2.INTER_AREA)
-        print("leaves resized...")
-    return leaf_image
-
-
+from custom_utils import *
 import random
 
 def calculate_leaf_location(cucumber_bbox, location='middle'):
@@ -128,21 +29,37 @@ def calculate_leaf_location(cucumber_bbox, location='middle'):
     else:
         raise ValueError(f"Invalid location option: {location}")
 
+def leaf_size_initialization(cucumber_mask, ratio = (0.4, 0.25)):
+    max_leaf_size_ratio, min_leaf_size_ratio = ratio  # 잎의 최대 및 최소 크기 비율
+    max_leaf_h, max_leaf_w = int(cucumber_mask.shape[0] * max_leaf_size_ratio), int(cucumber_mask.shape[1] * max_leaf_size_ratio)
+    min_leaf_h, min_leaf_w = int(cucumber_mask.shape[0] * min_leaf_size_ratio), int(cucumber_mask.shape[1] * min_leaf_size_ratio)
+
+    return max_leaf_h, max_leaf_w, min_leaf_h, min_leaf_w
 
 def resize_leaf_to_target_ratio(cucumber_mask, leaf_image, leaf_position, target_ratio):
+    loss_rate = 0.05  # 손실률
     # 잎의 중심 좌표
     leaf_x, leaf_y = leaf_position
-    cucumber_area = np.sum(cucumber_mask == 255)  # 오이의 총 면적 계산
+    cucumber_area = np.sum(cucumber_mask == 255)  # 오이 마스크 면적 계산
 
-    # 최소 및 최대 비율 제한
-    min_ratio, max_ratio = 0.01, 0.9
-    target_ratio = np.clip(target_ratio, min_ratio, max_ratio)
-    
-    # 잎의 초기 크기
+    # 초기 잎 크기 제한 (오이 이미지 크기 비율에 따라)
+    max_leaf_size_ratio, min_leaf_size_ratio = 0.4, 0.25  # 잎의 최대 크기를 오이 이미지의 50%로 제한
+    max_leaf_h, max_leaf_w, min_leaf_h, min_leaf_w = leaf_size_initialization(cucumber_mask, ratio=(max_leaf_size_ratio, min_leaf_size_ratio) )  # 잎 크기 초기화
+
+    # 잎 이미지 크기 제한
     leaf_h, leaf_w = leaf_image.shape[:2]
+    if leaf_h > max_leaf_h or leaf_w > max_leaf_w:
+        leaf_image = cv2.resize(leaf_image, (max_leaf_w, max_leaf_h), interpolation=cv2.INTER_LINEAR)
+    elif leaf_h < min_leaf_h or leaf_w < min_leaf_w:
+        leaf_image = cv2.resize(leaf_image, (min_leaf_w, min_leaf_h), interpolation=cv2.INTER_LINEAR)
 
+    leaf_h, leaf_w = leaf_image.shape[:2]
     start_y = max(0, leaf_y - leaf_h // 2)
     start_x = max(0, leaf_x - leaf_w // 2)
+
+    max_iterations = 200
+    iterations = 0
+
     while True:
         # 잎 마스크 생성 (알파 채널 활용)
         leaf_mask = (leaf_image[:, :, 3] > 0).astype(np.uint8) * 255
@@ -152,29 +69,42 @@ def resize_leaf_to_target_ratio(cucumber_mask, leaf_image, leaf_position, target
         end_y = min(cucumber_mask.shape[0], start_y + leaf_h)
         end_x = min(cucumber_mask.shape[1], start_x + leaf_w)
 
-        # 잎 마스크 복사
-        temp_leaf_mask[start_y:end_y, start_x:end_x] = leaf_mask[
-            0:end_y - start_y, 0:end_x - start_x
-        ]
+        temp_leaf_mask[start_y:end_y, start_x:end_x] = leaf_mask[0:end_y - start_y, 0:end_x - start_x]
 
         # 현재 겹침 영역 계산
         overlap_area = np.sum((temp_leaf_mask > 0) & (cucumber_mask > 0))
         current_ratio = overlap_area / cucumber_area
+        
+        # 디버깅 출력
+        print(f"Debug: Cucumber Area: {cucumber_area}, Overlap Area: {overlap_area}, Leaf Area: {leaf_h * leaf_w}, Current Ratio: {current_ratio:.4f}")
 
         # 목표 비율에 도달하면 종료
-        if abs(current_ratio - target_ratio) < 0.01:
+        if abs(current_ratio - target_ratio) < loss_rate and iterations > 0:
+            print(f"Target ratio achieved with current ratio: {current_ratio:.4f} after {iterations} iterations.")
             break
 
-        # 크기 조정 비율 계산
-        scale_factor = (target_ratio / current_ratio) ** 0.5
-        new_h = max(1, int(leaf_h * scale_factor))
-        new_w = max(1, int(leaf_w * scale_factor))
+        # 반복 초과 시 종료
+        if iterations >= max_iterations:
+            print(f"Error: Maximum iterations ({max_iterations}) reached. Current ratio: {current_ratio:.4f}. Exiting.")
+            exit(1)
 
+        # 크기 조정 비율 계산
+        if current_ratio < target_ratio:
+            scale_factor = min((target_ratio / current_ratio) ** 0.5, 1.02)
+        else:
+            scale_factor = max((target_ratio / current_ratio) ** 0.5, 0.98)
+        
+        
+        new_h, new_w = max(1, int(leaf_h * scale_factor)), max(1, int(leaf_w * scale_factor))
+        
         # 잎 이미지 크기 조정
         leaf_image = cv2.resize(leaf_image, (new_w, new_h), interpolation=cv2.INTER_LINEAR)
         leaf_h, leaf_w = leaf_image.shape[:2]  # 크기 업데이트
+        iterations += 1
 
+    print("Leaves resized to target ratio.")
     return leaf_image
+
 
 
 def merge_and_crop_leaf(cucumber_image, resized_leaf_image, leaf_position):
@@ -205,6 +135,11 @@ def merge_and_crop_leaf(cucumber_image, resized_leaf_image, leaf_position):
     cropped_h, cropped_w = cropped_leaf_image.shape[:2]
     for i in range(cropped_h):
         for j in range(cropped_w):
+            # 경계 초과 방지 조건
+            if (crop_y_start + i >= cucumber_image.shape[0]) or (crop_x_start + j >= cucumber_image.shape[1]):
+                print("====== Warning: Leaf image out of bounds. ======")
+                continue  # 경계를 벗어난 경우 스킵
+
             if cropped_leaf_image[i, j, 3] > 0:  # 투명하지 않은 경우
                 merged_image[crop_y_start + i, crop_x_start + j] = cropped_leaf_image[i, j, :3]
                 leaf_mask[crop_y_start + i, crop_x_start + j] = 255  # 잎 마스크 업데이트
@@ -212,5 +147,26 @@ def merge_and_crop_leaf(cucumber_image, resized_leaf_image, leaf_position):
     return merged_image, leaf_mask
 
 
+# modal 마스크 생성
+def generate_save_masks(cucumber_mask, leaf_mask, image_name, save_dir):
+    # 겹치는 부분 (오이 마스크와 잎 마스크가 동시에 255인 부분을 추출)
+    overlap_mask = (cucumber_mask == 255) & (leaf_mask == 255)
+    
+    # Modal 마스크 생성 (겹치는 부분을 제외한 오이 마스크)
+    modal_mask = cucumber_mask.copy()
+    modal_mask[overlap_mask] = 0  # overlap 영역을 0으로 만들어 겹친 부분 제거
 
+    # 마스크를 PNG 형식으로 저장 (이진화된 값 0 또는 255로 저장)
+    modal_filename = f"{os.path.splitext(image_name)[0]}_amodal_mask.png"
+    amodal_mask_path = save_image(save_dir, modal_filename, cucumber_mask)
+
+    # 마스크를 PNG 형식으로 저장 (이진화된 값 0 또는 255로 저장)
+    modal_filename = f"{os.path.splitext(image_name)[0]}_modal_mask.png"
+    modal_mask_path = save_image(save_dir, modal_filename, modal_mask)
+
+    # Overlap 마스크 저장
+    overlap_filename = f"{os.path.splitext(image_name)[0]}_overlap_mask.png"
+    overlap_mask_path = save_image(save_dir, overlap_filename, overlap_mask.astype(np.uint8) * 255)
+
+    return modal_mask, overlap_mask
